# docs/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CarePlus HMS API",
    "version": "1.0.0",
    "description": "Hospital Management REST API — patients, appointments, beds, pharmacy, lab, billing, doctors, departments, inventory, staff, audit and dashboard. All responses use { data, meta? } and errors use { error: { code, message, details } }."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Credential login (email + password)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "tokens"
          }
        },
        "security": []
      }
    },
...
    "/api/v1/auth/forgot-password": {
      "post": {
        "summary": "Get security question (enumeration-safe)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "question"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "summary": "Reset password via security answer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "answer": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "reset"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "summary": "Change own password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "changed"
          }
        }
      }
    },
...
  }
}